\name{close}
\alias{close}
\title{ Close a Database Object}
\description{
    See method documentation for details
}
\usage{
close(object, ...)
}
\arguments{
 \item{object}{ an object of class dbConnection or resultSet }
 \item{\dots}{ see method descriptions }
}
\references{\url{http://www.mysql.com/}}
\author{David James}
\seealso{ \code{\link{close.MySQLConnection}},
    \code{\link{close.MySQLResultSet}}}
\keyword{interface}
