% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.r
\docType{class}
\name{DB-class}
\alias{DB-class}
\alias{DB}
\title{Class object containing username, database and password for an MySQL query.}
\arguments{
\item{user}{Username for MySQL accession}

\item{database}{Database name which should be called in MySQL}

\item{password}{Password to access the database}
}
\description{
Class object containing username, database and password for an MySQL query.
}
\examples{
DB()
}
\keyword{Command}
\keyword{Line}
\keyword{MySQL}
\keyword{SQL}
