% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{db-meta}
\alias{db-meta}
\alias{dbGetException,MySQLConnection-method}
\alias{dbGetInfo,MySQLConnection-method}
\alias{dbListResults,MySQLConnection-method}
\alias{summary,MySQLConnection-method}
\title{Database interface meta-data}
\usage{
\S4method{dbGetInfo}{MySQLConnection}(dbObj, what = "", ...)

\S4method{dbListResults}{MySQLConnection}(conn, ...)

\S4method{summary}{MySQLConnection}(object, verbose = FALSE, ...)

\S4method{dbGetException}{MySQLConnection}(conn, ...)
}
\arguments{
\item{what}{optional}

\item{...}{Other arguments for compatibility with generic.}

\item{conn,dbObj,object}{MySQLConnection object.}
}
\description{
Database interface meta-data
}
\examples{
\dontrun{
con <- dbConnect(RMySQL::MySQL(), group = "wireless")
dbGetInfo(con)
dbListResults(con)
dbListTables(con)
}
}

