% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Extensions.R
\name{dbBuildTableDefinition}
\alias{dbBuildTableDefinition}
\title{Build the SQL CREATE TABLE definition as a string}
\usage{
dbBuildTableDefinition(dbObj, name, obj, field.types = NULL,
  row.names = TRUE, ...)
}
\arguments{
\item{dbObj}{any DBI object (used only to dispatch according to the engine
(e.g., MySQL, Oracle, PostgreSQL, SQLite)}

\item{name}{name of the new SQL table}

\item{obj}{an R object coerceable to data.frame for which we want to create
a table}

\item{field.types}{optional named list of the types for each field in
\code{obj}}

\item{row.names}{logical, should row.name of \code{value} be exported as a
\code{row\_names} field? Default is TRUE}

\item{\dots}{reserved for future use}
}
\value{
An SQL string
}
\description{
The output SQL statement is a simple \code{CREATE TABLE} with suitable for
\code{dbGetQuery}
}
\keyword{internal}

