% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{dbColumnInfo,MySQLConnection-method}
\alias{dbColumnInfo,MySQLConnection-method}
\title{Experimental dbColumnInfo method for a connection}
\usage{
\S4method{dbColumnInfo}{MySQLConnection}(res, name, ...)
}
\description{
Experimental dbColumnInfo method for a connection
}
\keyword{internal}

