% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transactions.R
\docType{methods}
\name{dbCommit,MySQLConnection-method}
\alias{dbCommit,MySQLConnection-method}
\alias{dbRollback,MySQLConnection-method}
\title{DBMS Transaction Management}
\usage{
\S4method{dbCommit}{MySQLConnection}(conn, ...)

\S4method{dbRollback}{MySQLConnection}(conn, ...)
}
\arguments{
\item{conn}{a \code{MySQLConnection} object, as produced by
\code{\link{dbConnect}}.}

\item{...}{Unused.}
}
\description{
Commits or roll backs the current transaction in an MySQL connection.
}

