% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{dbDataType,MySQLConnection,ANY-method}
\alias{dbDataType,MySQLConnection,ANY-method}
\alias{dbDataType,MySQLDriver,ANY-method}
\title{Determine the SQL Data Type of an S object}
\usage{
\S4method{dbDataType}{MySQLConnection,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{MySQLDriver,ANY}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A \code{MySQLDriver} or \code{MySQLConnection}.}

\item{obj}{R/S-Plus object whose SQL type we want to determine.}

\item{\dots}{any other parameters that individual methods may need.}
}
\description{
This method is a straight-forward implementation of the corresponding
generic function.
}
\examples{
library(DBI)
dbDataType(RMySQL::MySQL(), "a")
dbDataType(RMySQL::MySQL(), 1:3)
dbDataType(RMySQL::MySQL(), 2.5)
}

