% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MySQL.R
\docType{methods}
\name{dbDataType,MySQLObject-method}
\alias{dbDataType,MySQLObject-method}
\title{Determine the SQL Data Type of an S object}
\usage{
\S4method{dbDataType}{MySQLObject}(dbObj, obj)
}
\arguments{
\item{dbObj}{any \code{MySQLObject} object, e.g., \code{MySQLDriver},
\code{MySQLConnection}, \code{MySQLResult}.}

\item{obj}{R/S-Plus object whose SQL type we want to determine.}

\item{\dots}{any other parameters that individual methods may need.}
}
\description{
This method is a straight-forward implementation of the corresponding
generic function.
}
\examples{
dbDataType(RMySQL::MySQL(), "a")
dbDataType(RMySQL::MySQL(), 1:3)
dbDataType(RMySQL::MySQL(), 2.5)
}

