% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extension.R
\name{dbEscapeStrings}
\alias{dbEscapeStrings}
\title{Escape SQL-special characters in strings.}
\usage{
dbEscapeStrings(con, strings, ...)
}
\description{
DEPRECATED: Please use \code{dbQuoteString} instead.
}
\keyword{internal}

