% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbGetInfo-methods}
\alias{dbColumnInfo,MySQLConnection-method}
\alias{dbColumnInfo,MySQLResult-method}
\alias{dbColumnInfo-methods}
\alias{dbGetDBIVersion-methods}
\alias{dbGetInfo}
\alias{dbGetInfo,MySQLConnection-method}
\alias{dbGetInfo,MySQLDriver-method}
\alias{dbGetInfo,MySQLObject-method}
\alias{dbGetInfo,MySQLResult-method}
\alias{dbGetInfo-methods}
\alias{dbGetRowCount,MySQLResult-method}
\alias{dbGetRowCount-methods}
\alias{dbGetRowsAffected,MySQLResult-method}
\alias{dbGetRowsAffected-methods}
\alias{dbGetStatement,MySQLResult-method}
\alias{dbGetStatement-methods}
\alias{dbHasCompleted,MySQLResult-method}
\alias{dbHasCompleted-methods}
\title{Database interface meta-data}
\usage{
\S4method{dbGetInfo}{MySQLConnection}(dbObj, what = "", ...)
}
\description{
These methods are straight-forward implementations of the corresponding
generic functions.
}
\section{Methods}{
 \describe{ \item{dbObj}{ any object that implements some
functionality in the R/S-Plus interface to databases (a driver, a connection
or a result set).  } \item{res}{ an \code{MySQLResult}.}
\item{list()}{currently not being used.} }
}
\examples{
\dontrun{
drv <- dbDriver("MySQL")
con <- dbConnect(drv, group = "wireless")

dbListTables(con)

rs <- dbSendQuery(con, query.sql)
dbGetStatement(rs)
dbHasCompleted(rs)

info <- dbGetInfo(rs)
names(dbGetInfo(drv))

# DBIConnection info
names(dbGetInfo(con))

# DBIResult info
names(dbGetInfo(rs))
}
}
\references{
See the Database Interface definition document \code{DBI.pdf} in
the base directory of this package or
\url{http://stat.bell-labs.com/RS-DBI}.
}
\seealso{
\code{\link{MySQL}}, \code{\link[DBI]{dbDriver}},
\code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbSendQuery}},
\code{\link[DBI]{dbGetQuery}}, \code{\link[DBI]{fetch}},
\code{\link[DBI]{dbCommit}}, \code{\link[DBI]{dbGetInfo}},
\code{\link[DBI]{dbListTables}}, \code{\link[DBI]{dbReadTable}}.
}
\keyword{database}
\keyword{interface}
\keyword{methods}

