% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{dbGetInfo,MySQLDriver-method}
\alias{dbGetInfo,MySQLDriver-method}
\alias{dbListConnections,MySQLDriver-method}
\alias{summary,MySQLDriver-method}
\alias{show,MySQLDriver-method}
\title{Get information about a MySQL driver.}
\usage{
\S4method{dbGetInfo}{MySQLDriver}(dbObj, what = "", ...)

\S4method{dbListConnections}{MySQLDriver}(drv, ...)

\S4method{summary}{MySQLDriver}(object, verbose = FALSE, ...)

\S4method{show}{MySQLDriver}(object)
}
\arguments{
\item{dbObj, object, drv}{Object created by \code{\link{MySQL}}.}

\item{what}{Optional}

\item{...}{Ignored. Needed for compatibility with generic.}

\item{verbose}{If \code{TRUE}, print extra info.}
}
\description{
Get information about a MySQL driver.
}
\examples{
db <- RMySQL::MySQL()

db
dbGetInfo(db)
dbListConnections(db)
summary(db)
}
