% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbListTables-methods}
\alias{dbListConnections,MySQLDriver-method}
\alias{dbListConnections-methods}
\alias{dbListFields,MySQLConnection,character-method}
\alias{dbListFields,MySQLResult,missing-method}
\alias{dbListFields-methods}
\alias{dbListResults,MySQLConnection-method}
\alias{dbListResults-methods}
\alias{dbListTables,MySQLConnection-method}
\alias{dbListTables-methods}
\title{List items from an MySQL DBMS and from objects}
\usage{
\S4method{dbListFields}{MySQLConnection,character}(conn, name, ...)
}
\description{
These methods are straight-forward implementations of the corresponding
generic functions.
}
\section{Methods}{
 \describe{ \item{drv}{an \code{MySQLDriver}.}
\item{conn}{an \code{MySQLConnection}.} \item{name}{a character string with
the table name.} \item{list()}{currently not used.} }
}
\examples{
\dontrun{
drv <- dbDriver("MySQL")
# after working awhile...
for(con in dbListConnections(drv)){
   dbGetStatement(dbListResults(con))
}
}
}
\references{
See the Database Interface definition document \code{DBI.pdf} in
the base directory of this package or
\url{http://stat.bell-labs.com/RS-DBI}.
}
\seealso{
\code{\link{MySQL}}, \code{\link[DBI]{dbGetInfo}},
\code{\link[DBI]{dbColumnInfo}}, \code{\link[DBI]{dbDriver}},
\code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbSendQuery}}
}
\keyword{database}
\keyword{interface}
\keyword{methods}

