% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Extensions.R
\docType{methods}
\name{dbNextResult}
\alias{dbMoreResults}
\alias{dbMoreResults,MySQLConnection-method}
\alias{dbNextResult}
\alias{dbNextResult,MySQLConnection-method}
\title{Fetch next result set from an SQL script or stored procedure (experimental)}
\usage{
dbNextResult(con, ...)

\S4method{dbNextResult}{MySQLConnection}(con, ...)

dbMoreResults(con, ...)

\S4method{dbMoreResults}{MySQLConnection}(con, ...)
}
\arguments{
\item{con}{a connection object (see \code{\link[DBI]{dbConnect}}).}

\item{...}{any additional arguments to be passed to the dispatched method}
}
\value{
\code{dbNextResult} returns a result set or \code{NULL}.

  \code{dbMoreResults} returns a logical specifying whether or not there are
  additional result sets to process in the connection.
}
\description{
SQL scripts (i.e., multiple SQL statements separated by ';') and stored
procedures oftentimes generate multiple result sets.  These generic
functions provide a means to process them sequentially. \code{dbNextResult}
fetches the next result from the sequence of pending results sets;
\code{dbMoreResults} returns a logical to indicate whether there are
additional results to process.
}
\examples{
\dontrun{
con <- dbConnect(MySQL(),
          dbname = "rs-dbi",
          client.flag=CLIENT_MULTI_STATEMENTS)
sql.script <- paste(
   "select * from abc",
   "select * def",
   collapse = ";")

rs1 <- dbSendQuery(con, sql.script)
data1 <- fetch(rs1, n = -1)

if(dbMoreResults(con)){
   rs2 <- dbNextResult(con)
   ## you could use dbHasCompleted(rs2) to determine whether
   ## rs2 is a select-like that generates output or not.
   data2 <- fetch(rs2, n = -1)
   }
}
}

