% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbSendQuery-methods}
\alias{dbClearResult,MySQLResult-method}
\alias{dbClearResult-methods}
\alias{dbGetException,MySQLConnection-method}
\alias{dbGetException,MySQLResult-method}
\alias{dbGetException-methods}
\alias{dbGetQuery,MySQLConnection,character-method}
\alias{dbGetQuery-methods}
\alias{dbSendQuery,MySQLConnection,character-method}
\alias{dbSendQuery-methods}
\title{Execute a statement on a given database connection}
\usage{
\S4method{dbSendQuery}{MySQLConnection,character}(conn, statement)
}
\description{
These methods are straight-forward implementations of the corresponding
generic functions.
}
\section{Methods}{
 \describe{ \item{conn}{ an \code{MySQLConnection} object.
} \item{statement}{a character vector of length 1 with the SQL statement.}
\item{res}{an \code{MySQLResult} object.} \item{list()}{additional
parameters.}\item{ }{additional parameters.} }
}
\examples{
\dontrun{
drv <- dbDriver("MySQL")
con <- dbConnect(drv, "usr", "password", "dbname")
res <- dbSendQuery(con, "SELECT * from liv25")
data <- fetch(res, n = -1)
}
}
\references{
See the Database Interface definition document \code{DBI.pdf} in
the base directory of this package or
\url{http://stat.bell-labs.com/RS-DBI}.
}
\seealso{
\code{\link{MySQL}}, \code{\link[DBI]{dbDriver}},
\code{\link[DBI]{dbConnect}}, \code{\link[DBI]{fetch}},
\code{\link[DBI]{dbCommit}}, \code{\link[DBI]{dbGetInfo}},
\code{\link[DBI]{dbReadTable}}.
}
\keyword{database}
\keyword{interface}
\keyword{methods}

