% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Driver.R
\docType{methods}
\name{dbUnloadDriver,MySQLDriver-method}
\alias{dbUnloadDriver,MySQLDriver-method}
\title{Unload MySQL driver.}
\usage{
\S4method{dbUnloadDriver}{MySQLDriver}(drv, ...)
}
\arguments{
\item{drv}{Object created by \code{\link{MySQL}}.}

\item{...}{Ignored. Needed for compatibility with generic.}
}
\value{
A logical indicating whether the operation succeeded or not.
}
\description{
Unload MySQL driver.
}

