\name{getException}
\alias{getException}
\title{ Get Exceptions Associated With a SQL Driver Object}
\description{
    For a MySQLConnection or MySQLResultSet object, get associated
    exceptions.
}
\usage{
getException(object)
}
\arguments{
 \item{object}{ object of class MySQLConnection or MySQLResultSet}
}
\references{\url{http://www.mysql.com/}}
\author{David James}
\examples{
}
\keyword{interface}
