% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MySQL.R
\docType{methods}
\name{make.db.names-methods}
\alias{SQLKeywords,MySQLObject-method}
\alias{SQLKeywords,missing-method}
\alias{SQLKeywords-methods}
\alias{isSQLKeyword,MySQLObject,character-method}
\alias{isSQLKeyword-methods}
\alias{make.db.names,MySQLObject,character-method}
\alias{make.db.names-methods}
\title{Make R/S-Plus identifiers into legal SQL identifiers}
\usage{
\S4method{make.db.names}{MySQLObject,character}(dbObj, snames,
  keywords = .MySQLKeywords, unique = TRUE, allow.keywords = TRUE, ...)
}
\description{
These methods are straight-forward implementations of the corresponding
generic functions.
}
\section{Methods}{
 \describe{ \item{dbObj}{ any MySQL object (e.g.,
\code{MySQLDriver}).  } \item{snames}{ a character vector of R/S-Plus
identifiers (symbols) from which we need to make SQL identifiers.  }
\item{name}{ a character vector of SQL identifiers we want to check against
keywords from the DBMS. } \item{unique}{ logical describing whether the
resulting set of SQL names should be unique.  Its default is \code{TRUE}.
Following the SQL 92 standard, uniqueness of SQL identifiers is determined
regardless of whether letters are upper or lower case.  }
\item{allow.keywords }{ logical describing whether SQL keywords should be
allowed in the resulting set of SQL names.  Its default is \code{TRUE} }
\item{keywords}{ a character vector with SQL keywords, by default it is
\code{.MySQLKeywords} define in \code{RMySQL}. This may be overriden by
users.  } \item{case}{ a character string specifying whether to make the
comparison as lower case, upper case, or any of the two.  it defaults to
\code{any}.  } \item{list()}{currently not used.} }
}
\examples{
\dontrun{
# This example shows how we could export a bunch of data.frames
# into tables on a remote database.

con <- dbConnect("MySQL", "user", "password")

export <- c("trantime.email", "trantime.print", "round.trip.time.email")
tabs <- make.db.names(export, unique = T, allow.keywords = T)

for(i in seq(along = export) )
   dbWriteTable(con, name = tabs[i],  get(export[i]))
}
}
\references{
The set of SQL keywords is stored in the character vector
\code{.SQL92Keywords} and reflects the SQL ANSI/ISO standard as documented
in "X/Open SQL and RDA", 1994, ISBN 1-872630-68-8.  Users can easily
override or update this vector.

MySQL does add some keywords to the SQL 92 standard, they are listed in the
\code{.MySQLKeywords} object.

See the Database Interface definition document \code{DBI.pdf} in the base
directory of this package or \url{http://stat.bell-labs.com/RS-DBI}.
}
\seealso{
\code{\link{MySQL}}, \code{\link[DBI]{dbReadTable}},
\code{\link[DBI]{dbWriteTable}}, \code{\link[DBI]{dbExistsTable}},
\code{\link[DBI]{dbRemoveTable}}, \code{\link[DBI]{dbListTables}}.
}
\keyword{database}
\keyword{interface}
\keyword{methods}

