% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{make.db.names,MySQLConnection,character-method}
\alias{make.db.names,MySQLConnection,character-method}
\alias{SQLKeywords,MySQLConnection-method}
\alias{isSQLKeyword,MySQLConnection,character-method}
\title{Make R/S-Plus identifiers into legal SQL identifiers}
\usage{
\S4method{make.db.names}{MySQLConnection,character}(dbObj, snames,
  keywords = .SQL92Keywords, unique = TRUE, allow.keywords = TRUE, ...)

\S4method{SQLKeywords}{MySQLConnection}(dbObj, ...)

\S4method{isSQLKeyword}{MySQLConnection,character}(dbObj, name,
  keywords = .MySQLKeywords, case = c("lower", "upper", "any")[3], ...)
}
\description{
DEPRECATED: please use \code{dbQuoteIdentifier} instead.
}
\keyword{internal}
