% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quote.R
\name{mysql-quoting}
\alias{mysql-quoting}
\title{Quote MySQL strings and identifiers.}
\description{
In MySQL, identifiers are enclosed in backticks, e.g. \code{`x`}.
}
\examples{
if (mysqlHasDefault()) {
  con <- dbConnect(RMySQL::MySQL())
  dbQuoteIdentifier(con, c("a b", "a`b"))
  dbQuoteString(con, c("a b", "a'b"))
  dbDisconnect(con)
}
}
\keyword{internal}

