% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mysqlBuildTableDefinition}
\alias{mysqlBuildTableDefinition}
\title{Build the SQL CREATE TABLE definition as a string}
\usage{
mysqlBuildTableDefinition(dbObj, name, obj, field.types = NULL,
  row.names = TRUE, ...)
}
\description{
DEPRECATED
}
\keyword{internal}
