\name{mysqlClientLibraryVersions}
\alias{mysqlClientLibraryVersions}
\title{MySQL Check for Compiled Versus Loaded Client Library Versions}
\description{
	This function prints out the compiled and loaded client library versions.
}
\usage{
mysqlClientLibraryVersions()
}
\value{
  A named integer vector of length two, the first element representing the 
  compiled library version and the second element representint the loaded client
  library version.
}
\keyword{database}
% vim: syntax=tex
