% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/default.R
\name{mysqlDefault}
\alias{mysqlDefault}
\alias{mysqlHasDefault}
\title{Connect to a default database.}
\usage{
mysqlDefault()

mysqlHasDefault()
}
\description{
Connect to a database configured by environment variables. The default
database name is "test", override with env var \code{MYSQL_DATABASE}.
Specify default username and password in \code{MYSQL_USER} and
\code{MYSQL_PASSWORD} envvars; or alternatively configured in
\code{my.ini} or \code{my.cnf}.
}
\examples{
if (mysqlHasDefault()) {
  db <- mysqlDefault()
  dbListTables(db)
  dbDisconnect(db)
}
}

