% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{mysqlExecQuery}
\alias{mysqlExecQuery}
\title{Execute a query on the server (no binding).}
\usage{
mysqlExecQuery(con, sql)
}
\arguments{
\item{con}{A MySQL connection.}

\item{sql}{A sql string to execute}
}
\description{
MySQL has two APIs for submitting queries, one for parameterised queries
and one for unparameterised. In most cases, you can use the parameterised
query interface even if you have zero parameter. However, some queries
(e.g. transaction modification) can not be executed through the
paramaterised interface. This function allows you to submit those queries
directly.
}
