% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/result.R
\docType{class}
\name{MySQLResult-class}
\alias{MySQLResult-class}
\alias{show,MySQLResult-method}
\title{Class MySQLResult}
\usage{
\S4method{show}{MySQLResult}(object)
}
\description{
MySQL's query results class.  This classes encapsulates the result of an SQL
statement (either \code{select} or not).
}
\keyword{internal}

