% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMySQL.R
\docType{package}
\name{RMySQL}
\alias{RMySQL}
\alias{RMySQL-package}
\alias{RMySQL-package}
\title{RMySQL: Database Interface and MySQL Driver for R}
\description{
Implements a DBI interface to MySQL and MariaDB databases.
}
\seealso{
the \href{https://cran.r-project.org/web/packages/RMySQL/RMySQL.pdf}{package documentation}.
}
