% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.r
\name{sql_query,DB,character-method}
\alias{sql_query,DB,character-method}
\title{Executes an SQL query on the command line}
\usage{
\S4method{sql_query}{DB,character}(object, query)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function executes an defined SQL query on the command line.
The
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
