% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MySQL.R
\docType{methods}
\name{summary-methods}
\alias{coerce,MySQLConnection,MySQLDriver-method}
\alias{coerce,MySQLConnection,MySQLResult-method}
\alias{coerce,MySQLObject,MySQLDriver-method}
\alias{coerce,MySQLObject,character-method}
\alias{coerce,MySQLObject,integer-method}
\alias{coerce,MySQLObject,numeric-method}
\alias{coerce,MySQLResult,MySQLConnection-method}
\alias{coerce,MySQLResult,MySQLDriver-method}
\alias{coerce,dbObjectId,integer-method}
\alias{coerce-methods}
\alias{format,MySQLObject-method}
\alias{format-methods}
\alias{print,MySQLObject-method}
\alias{show,MySQLObject-method}
\alias{show-methods}
\alias{summary,MySQLConnection-method}
\alias{summary,MySQLDriver-method}
\alias{summary,MySQLObject-method}
\alias{summary,MySQLResult-method}
\alias{summary-methods}
\title{Summarize an MySQL object}
\description{
These methods are straight-forward implementations of the corresponding
generic functions.
}
\section{Methods}{
 \describe{

\item{object = "DBIObject"}{ Provides relevant metadata information on
\code{object}, for instance, the MySQL server file, the SQL statement
associated with a result set, etc.  } \item{from}{object to be coerced}
\item{to}{coercion class} \item{x}{object to \code{format} or \code{print}
or \code{show}} }
}
\keyword{database}
\keyword{interface}
\keyword{methods}

