\name{supportObjects}
\alias{.SQLKeywords}
\alias{.conflicts.OK}
\non_function{}
\title{Support objects for RS-DBI}
\usage{data(.SQLKeywords)}
\description{
  .SQLKeywords is a vector of SQL92 reserved words -- it may be
  overriden to include additional identifiers reserved by 
  RDBMS.
}
\keyword{dataset}
