% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tables.R
\docType{methods}
\name{table}
\alias{dbExistsTable,MySQLConnection,character-method}
\alias{dbListFields,MySQLConnection,character-method}
\alias{dbListTables,MySQLConnection-method}
\alias{dbReadTable,MySQLConnection,character-method}
\alias{dbRemoveTable,MySQLConnection,character-method}
\alias{dbWriteTable,MySQLConnection,character,character-method}
\alias{dbWriteTable,MySQLConnection,character,data.frame-method}
\alias{table}
\title{Convenience functions for importing/exporting DBMS tables}
\usage{
\S4method{dbReadTable}{MySQLConnection,character}(conn, name,
  row.names = "row_names", check.names = TRUE, ...)

\S4method{dbWriteTable}{MySQLConnection,character,data.frame}(conn, name, value,
  field.types, row.names = TRUE, overwrite = FALSE, append = FALSE, ...,
  allow.keywords = FALSE)

\S4method{dbWriteTable}{MySQLConnection,character,character}(conn, name, value,
  field.types = NULL, overwrite = FALSE, append = FALSE, header,
  row.names, nrows = 50, sep = ",", eol = "\\n", skip = 0,
  quote = "\\"", ...)

\S4method{dbListTables}{MySQLConnection}(conn, ...)

\S4method{dbExistsTable}{MySQLConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{MySQLConnection,character}(conn, name, ...)

\S4method{dbListFields}{MySQLConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{a \code{\linkS4class{SQLiteConnection}} object, produced by
\code{\link[DBI]{dbConnect}}}

\item{name}{a character string specifying a table name. SQLite table names
are \emph{not} case sensitive, e.g., table names \code{ABC} and \code{abc}
are considered equal.}

\item{row.names}{A string or an index specifying the column in the DBMS table
to use as \code{row.names} in the output data.frame. Defaults to using the
\code{row_names} column if present. Set to \code{NULL} to never use
row names.}

\item{check.names}{If \code{TRUE}, the default, column names will be
converted to valid R identifiers.}
}
\value{
A data.frame in the case of \code{dbReadTable}; otherwise a logical
indicating whether the operation was successful.
}
\description{
These functions mimic their R/S-Plus counterpart \code{get}, \code{assign},
\code{exists}, \code{remove}, and \code{objects}, except that they generate
code that gets remotely executed in a database engine.
}
\note{
Note that the data.frame returned by \code{dbReadTable} only has
primitive data, e.g., it does not coerce character data to factors.
}

