% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{constants}
\alias{.MySQLPkgName}
\alias{.MySQLPkgRCS}
\alias{.MySQLPkgVersion}
\alias{.MySQLSQLKeywords}
\alias{CLIENT_COMPRESS}
\alias{CLIENT_CONNECT_WITH_DB}
\alias{CLIENT_FOUND_ROWS}
\alias{CLIENT_IGNORE_SIGPIPE}
\alias{CLIENT_IGNORE_SPACE}
\alias{CLIENT_INTERACTIVE}
\alias{CLIENT_LOCAL_FILES}
\alias{CLIENT_LONG_FLAG}
\alias{CLIENT_LONG_PASSWORD}
\alias{CLIENT_MULTI_RESULTS}
\alias{CLIENT_MULTI_STATEMENTS}
\alias{CLIENT_NO_SCHEMA}
\alias{CLIENT_ODBC}
\alias{CLIENT_PROTOCOL_41}
\alias{CLIENT_RESERVED}
\alias{CLIENT_SECURE_CONNECTION}
\alias{CLIENT_SSL}
\alias{CLIENT_TRANSACTIONS}
\alias{constants}
\title{Constants}
\description{
Constants
}
\section{Constants}{
 \code{.MySQLPkgName} (currently \code{"RMySQL"}),
\code{.MySQLPkgVersion} (the R package version), \code{.MySQLPkgRCS} (the
RCS revision), \code{.MySQLSQLKeywords} (a lot!)
}

