% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connection.R, R/zzz_compatibility.R
\docType{methods}
\name{db-meta}
\alias{db-meta}
\alias{dbGetInfo,MySQLConnection-method}
\alias{dbObjectId-class}
\alias{show,MySQLConnection-method}
\title{Database interface meta-data}
\usage{
\S4method{dbGetInfo}{MySQLConnection}(dbObj, what = "", ...)

\S4method{show}{MySQLConnection}(object)
}
\arguments{
\item{what}{optional}

\item{...}{Other arguments for compatibility with generic.}

\item{conn,dbObj,object}{MySQLConnection object.}
}
\description{
Database interface meta-data
}
\examples{
if (mysqlHasDefault()) {
  con <- dbConnect(RMySQL::MySQL(), dbname = "test")

  summary(con)

  dbGetInfo(con)
  dbListResults(con)
  dbListTables(con)

  dbDisconnect(con)
}
}

