% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Extensions.R
\docType{methods}
\name{dbApply}
\alias{dbApply}
\alias{dbApply,MySQLResult-method}
\title{Apply R/S-Plus functions to remote groups of DBMS rows (experimental)}
\usage{
dbApply(res, ...)

\S4method{dbApply}{MySQLResult}(res, ...)
}
\description{
Applies R/S-Plus functions to groups of remote DBMS rows without bringing an
entire result set all at once.  The result set is expected to be sorted by
the grouping field.
}
\section{Methods}{
 \describe{ \item{res}{a MySQL result set (see
\code{\link[DBI]{dbSendQuery}}).} \item{...}{any additional arguments to be
passed to \code{FUN}.} }
}
\examples{
\dontrun{
## compute quanitiles for each network agent
con <- dbConnect(MySQL(), group="vitalAnalysis")
rs <- dbSendQuery(con,
             "select Agent, ip_addr, DATA from pseudo_data order by Agent")
out <- dbApply(rs, INDEX = "Agent",
        FUN = function(x, grp) quantile(x$DATA, names=FALSE))
}
}
\references{
See the Database Interface definition document \code{DBI.pdf} in
the base directory of this package or
\url{http://stat.bell-labs.com/RS-DBI}.
}
\seealso{
\code{\link{MySQL}} \code{\link{mysqlDBApply}}
\code{\link[DBI]{dbSendQuery}} \code{\link[DBI]{fetch}}
}
\keyword{database}
\keyword{interface}
\keyword{programming}

