% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{dbColumnInfo,MySQLConnection-method}
\alias{dbColumnInfo,MySQLConnection-method}
\title{Experimental dbColumnInfo method for a connection}
\usage{
\S4method{dbColumnInfo}{MySQLConnection}(res, name, ...)
}
\description{
Experimental dbColumnInfo method for a connection
}
\keyword{internal}
