% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transactions.R
\docType{methods}
\name{dbCommit,MySQLConnection-method}
\alias{dbCommit,MySQLConnection-method}
\title{DBMS Transaction Management}
\usage{
\S4method{dbCommit}{MySQLConnection}(conn, ...)
}
\description{
Commits or roll backs the current transaction in an MySQL connection
}
\section{Methods}{
 \describe{ \item{conn}{a \code{MySQLConnection} object, as
produced by the function \code{dbConnect}.} \item{list()}{currently
unused.}\item{ }{currently unused.} }
}
\examples{
\dontrun{
drv <- dbDriver("MySQL")
con <- dbConnect(drv, group = "group")
rs <- dbSendQuery(con,
      "delete * from PURGE as p where p.wavelength<0.03")
if(dbGetInfo(rs, what = "rowsAffected") > 250){
  warning("dubious deletion -- rolling back transaction")
  dbRollback(con)
}
}
}
\references{
See the Database Interface definition document \code{DBI.pdf} in
the base directory of this package or
\url{http://stat.bell-labs.com/RS-DBI}.
}
\seealso{
\code{\link{MySQL}}, \code{\link[DBI]{dbConnect}},
\code{\link[DBI]{dbSendQuery}}, \code{\link[DBI]{dbGetQuery}},
\code{\link[DBI]{fetch}}, \code{\link[DBI]{dbCommit}},
\code{\link[DBI]{dbGetInfo}}, \code{\link[DBI]{dbReadTable}}.
}
\keyword{database}
\keyword{interface}
\keyword{methods}

