% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MySQL.R
\docType{methods}
\name{dbDataType,MySQLObject-method}
\alias{dbDataType,MySQLObject-method}
\title{Determine the SQL Data Type of an S object}
\usage{
\S4method{dbDataType}{MySQLObject}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{any \code{MySQLObject} object, e.g., \code{MySQLDriver},
\code{MySQLConnection}, \code{MySQLResult}.}

\item{obj}{R/S-Plus object whose SQL type we want to determine.}

\item{\dots}{any other parameters that individual methods may need.}
}
\description{
This method is a straight-forward implementation of the corresponding
generic function.
}
\examples{
\dontrun{
data(quakes)
drv <- dbDriver("MySQL")
sql.type <- dbDataType(drv, quakes)
}
}
\references{
See the Database Interface definition document \code{DBI.pdf} in
the base directory of this package or
\url{http://stat.bell-labs.com/RS-DBI}.
}
\seealso{
\code{\link[DBI]{isSQLKeyword}} \code{\link[DBI]{make.db.names}}
}
\keyword{database}
\keyword{interface}
\keyword{methods}

