% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dbEscapeStrings}
\alias{dbEscapeStrings}
\title{Escape SQL-special characters in strings.}
\usage{
dbEscapeStrings(con, strings, ...)
}
\description{
DEPRECATED: Please use \code{dbQuoteString} instead.
}
\keyword{internal}

