% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extension.R
\docType{methods}
\name{dbQuoteIdentifier,MySQLConnection,character-method}
\alias{dbQuoteIdentifier,MySQLConnection,character-method}
\title{Quote method for MySQL identifiers}
\usage{
\S4method{dbQuoteIdentifier}{MySQLConnection,character}(conn, x, ...)
}
\description{
In MySQL, identifiers are enclosed in backticks, e.g. \code{`x`}.
}
\keyword{internal}

