% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{isIdCurrent}
\alias{isIdCurrent}
\title{Check if a database object is valid.}
\usage{
isIdCurrent(obj)
}
\value{
a logical scalar.
}
\description{
DEPRECATED
}
\examples{
library(DBI)
dbIsValid(MySQL())
}
\keyword{internal}

