% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MySQL.R
\name{isIdCurrent}
\alias{isIdCurrent}
\title{Check whether a database handle object is valid or not}
\usage{
isIdCurrent(obj)
}
\arguments{
\item{obj}{any \code{dbObject} (e.g., \code{dbDriver}, \code{dbConnection},
\code{dbResult}).}
}
\value{
a logical scalar.
}
\description{
Support function that verifies that an object holding a reference to a
foreign object is still valid for communicating with the RDBMS
}
\details{
\code{dbObjects} are R/S-Plus remote references to foreign objects. This
introduces differences to the object's semantics such as persistence (e.g.,
connections may be closed unexpectedly), thus this function provides a
minimal verification to ensure that the foreign object being referenced can
be contacted.
}
\examples{
\dontrun{
cursor <- dbSendQuery(con, sql.statement)
isIdCurrent(cursor)
}
}
\seealso{
\code{\link[DBI]{dbDriver}} \code{\link[DBI]{dbConnect}}
\code{\link[DBI]{dbSendQuery}} \code{\link[DBI]{fetch}}
}
\keyword{database}
\keyword{interface}

