% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\docType{methods}
\name{mysql-quoting}
\alias{dbQuoteIdentifier,MySQLConnection,character-method}
\alias{dbQuoteString,MySQLConnection,character-method}
\alias{mysql-quoting}
\title{Quote MySQL strings and identifiers.}
\usage{
\S4method{dbQuoteIdentifier}{MySQLConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{MySQLConnection,character}(conn, x, ...)
}
\description{
In MySQL, identifiers are enclosed in backticks, e.g. \code{`x`}.
}
\examples{
if (mysqlHasDefault()) {
  con <- dbConnect(RMySQL::MySQL())
  dbQuoteIdentifier(con, c("a b", "a`b"))
  dbQuoteString(con, c("a b", "a'b"))
  dbDisconnect(con)
}
}
\keyword{internal}

