% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MySQLDriver.R
\name{mysqlClientLibraryVersions}
\alias{mysqlClientLibraryVersions}
\title{MySQL Check for Compiled Versus Loaded Client Library Versions}
\usage{
mysqlClientLibraryVersions()
}
\value{
A named integer vector of length two, the first element representing
the compiled library version and the second element representint the loaded
client library version.
}
\description{
This function prints out the compiled and loaded client library versions.
}
\examples{
mysqlClientLibraryVersions()
}

