% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MySQLConnection.R
\docType{class}
\name{MySQLConnection-class}
\alias{MySQLConnection-class}
\title{Class MySQLConnection.}
\description{
\code{MySQLConnection.} objects are usually created by
\code{\link[DBI]{dbConnect}}
}
\keyword{internal}

