% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MySQLDriver.R
\docType{class}
\name{MySQLDriver-class}
\alias{MySQLDriver-class}
\alias{dbUnloadDriver,MySQLDriver-method}
\alias{dbIsValid,MySQLDriver-method}
\title{Class MySQLDriver with constructor MySQL.}
\usage{
\S4method{dbUnloadDriver}{MySQLDriver}(drv, ...)

\S4method{dbIsValid}{MySQLDriver}(dbObj)
}
\description{
An MySQL driver implementing the R database (DBI) API.
This class should always be initialized with the \code{MySQL()} function.
It returns a singleton that allows you to connect to MySQL.
}
\keyword{internal}
