% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/result.R
\docType{class}
\name{MySQLResult-class}
\alias{MySQLResult-class}
\title{Class MySQLResult}
\description{
MySQL's query results class.  This classes encapsulates the result of an SQL
statement (either \code{select} or not).
}
\keyword{internal}

