% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Result.R
\docType{class}
\name{MySQLResult-class}
\alias{MySQLResult-class}
\title{Class MySQLResult}
\description{
MySQL's query results class.  This classes encapsulates the result of an SQL
statement (either \code{select} or not).
}
\section{Generators}{
 The main generator is \code{\link[DBI]{dbSendQuery}}.
}
\examples{
\dontrun{
drv <- dbDriver("MySQL")
con <- dbConnect(drv, dbname = "rsdbi.db")
}
}
\references{
See the Database Interface definition document \code{DBI.pdf} in
the base directory of this package or
\url{http://developer.r-project.org/db}.
}
\seealso{
DBI base classes:

\code{\link[DBI]{DBIObject-class}} \code{\link[DBI]{DBIDriver-class}}
\code{\link[DBI]{DBIConnection-class}} \code{\link[DBI]{DBIResult-class}}

MySQL classes:

\code{\link{MySQLObject-class}} \code{\link{MySQLDriver-class}}
\code{\link{MySQLConnection-class}} \code{\link{MySQLResult-class}}
}
\keyword{classes}
\keyword{database}
\keyword{interface}

