% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MySQLResult.R
\docType{class}
\name{MySQLResult-class}
\alias{MySQLResult-class}
\alias{dbIsValid,MySQLResult-method}
\title{Class MySQLResult}
\usage{
\S4method{dbIsValid}{MySQLResult}(dbObj)
}
\description{
MySQL's query results class.  This classes encapsulates the result of an SQL
statement (either \code{select} or not).
}
\keyword{internal}

