% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{methods}
\name{result-meta}
\alias{dbColumnInfo,MySQLResult-method}
\alias{dbGetRowCount,MySQLResult-method}
\alias{dbGetRowsAffected,MySQLResult-method}
\alias{dbHasCompleted,MySQLResult-method}
\alias{result-meta}
\title{Database interface meta-data.}
\usage{
\S4method{dbColumnInfo}{MySQLResult}(res, ...)

\S4method{dbGetRowsAffected}{MySQLResult}(res, ...)

\S4method{dbGetRowCount}{MySQLResult}(res, ...)

\S4method{dbHasCompleted}{MySQLResult}(res, ...)
}
\arguments{
\item{res}{An object of class \code{\linkS4class{MySQLResult}}}

\item{...}{Ignored. Needed for compatibility with generic}
}
\description{
See documentation of generics for more details.
}
\examples{
if (mysqlHasDefault()) {
con <- dbConnect(RMySQL::MySQL(), dbname = "test")
dbWriteTable(con, "t1", datasets::USArrests, overwrite = TRUE)

rs <- dbSendQuery(con, "SELECT * FROM t1 WHERE UrbanPop >= 80")
rs

dbGetStatement(rs)
dbHasCompleted(rs)
dbColumnInfo(rs)

dbFetch(rs)
rs

dbClearResult(rs)
dbRemoveTable(con, "t1")
dbDisconnect(con)
}
}

