% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMySQL.R
\docType{package}
\name{RMySQL-package}
\alias{RMySQL}
\alias{RMySQL-package}
\title{RMySQL: Database Interface and MySQL Driver for R}
\description{
Implements DBI-compliant Interface to MySQL and MariaDB Databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstats-db/rmysql}
  \item \url{https://downloads.mariadb.org/connector-c/}
  \item Report bugs at \url{https://github.com/rstats-db/rmysql/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu}

Authors:
\itemize{
  \item David James
  \item Saikat DebRoy
  \item Hadley Wickham \email{hadley@rstudio.com}
  \item Jeffrey Horner
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item Kungliga Tekniska Högskolan (Source code for timegm) [contributor]
}

}
