\name{S4R}
\alias{as}
\alias{new}
\alias{oldClass}
\alias{oldClass<-}
\alias{show}
\alias{usingR}

\title{R compatibility with S version 4/Splus5 support functions}
\description{
  These functions ease the task of porting functions into R
  from S Version 4 and S-Plus 5.0 and later.  See the documentation
  there.
}
\usage{
as(object, classname)
new("class", ...)
oldClass(x)
oldClass(x) <- "class"
show(x)
usingR(major, minor)
}
\keyword{compatibility}
